+++ src/utils/round.js
@@ -1,6 +1,9 @@
 export function round(n, precision = 2) {
   if ("number" === typeof n) {
-    return n.toFixed(precision).replace(".00", "");
+    return n
+      .toFixed(precision)
+      .replace(/([0-9]*\.[1-9]*)(0*)$/, "$1")
+      .replace(/\.$/, "");
   }
   return n;
 }